#33 Move decorators to utils module
