--- decode/core/__init__.py
@@ -1,2 +1,0 @@
-# dependencies
-from .array.decorators import *
